==> notebooks/data_preprocessing/00_intro.ipynb <==
# Data and file structure

The available raw data for midatams is stored in the data folder. There are two directories of interest:

- data/seantis: contains reports (text) and labels for seantis data in csv format. The data in this folder will be used to train various classifiers.
- data/reports_with_struct_data: this data (rsd) will be labelled with the trained classifiers. There is some overlap between these reports and the seantis reports.

The data was imported from the hospital at various stages in time, thus the `imported_time` subdirectories in `data/seantis` and `data/reports_with_struct_data`. Of interest is only the newest one `imported_20210507`. To match reports to labels I will use the `research_id` entry provided in the datafiles. This id is unique to a single patient.

In the rest of this notebook I provide a short overview of the tasks I want to perform on the data as well as the data files I will be using for this. The tasks were defined by a previous project, the goal of this thesis is to reimplement these tasks with more modern approaches.

## 01 Classifying text lines

A classifier is trained to classify single text lines of a report. For example, the lines can be classified as containing a diagnosis (`"dm"`) or history (`his`) of a patient. For all classes see the file `01_classifying_text_lines`. This task was used as a preprocessing step to later steps of structured information extraction. So that only lines classified as containing a diagnosis will be fed to a downstream classifier, extracting the exact diagnosis. This step might be unnecessary with modern transformers that can handle longer text inputs.

## 02 Classifiyng MS-Diagnosis



import sys
import os
sys.path.append(os.getcwd()+"/../..")
from src import paths
import pandas as pd
import importlib
from src.data.data_loader import get_nested_csv

# Kisim Diagnoses
# Extracting the data
df = get_nested_csv(dir_name="seantis", file_name="kisim_diagnoses.csv")

# Saving the data
df.to_csv(os.path.join(paths.DATA_PATH_PREPROCESSED, "kisim_diagnoses_extracted.csv"), index=False)

# Diagnoses
# Extracting the data
df = get_nested_csv(dir_name="seantis", file_name="diagnoses.csv")

# Saving the data
df.to_csv(os.path.join(paths.DATA_PATH_PREPROCESSED, "diagnoses_extracted.csv"), index=False)
